Create constants.py and launcher.py
- Relocate the constants (`OUTPUT_FORMATS`, `DEFAULT_PROGRAMS`, etc.)
  to `src/pydot/constants.py`, and import where needed in other code
- Relocate `call_graphviz` and its support functions to
  `src/pydot/launcher.py`, importing where needed (`core.py`)
